Add SendSMS Functionality Modified SendEmail method to send customized message Modified CreatStudentServlet to email student and parent for password update upon account creation Debug CreateStudent and Delete Student Functionality -insert ignore if the parent exists in database -show error message if duplicated student detect -didn't delete parent if that parent has more than 1 child

+++ SteppingStones/src/main/java/controller/DeleteStudentServlet.java
@@ -44,7 +44,12 @@ public class DeleteStudentServlet extends HttpServlet {
         int parentID = ParentChildRelDAO.getParentID(studentID);
         boolean deleteStudent = StudentDAO.deleteStudentbyID(studentID);
         boolean deleteParentChildRel = ParentChildRelDAO.deleteParentChildRel(studentID);
-        boolean deleteParent = ParentDAO.deleteParent(parentID);
+        boolean deleteParent;
+        if(ParentChildRelDAO.getNumOfChild(parentID) == 1){
+            deleteParent = ParentDAO.deleteParent(parentID);
+        }else{
+            deleteParent = true;
+        }     
         boolean deleteTuitionGrade = StudentGradeDAO.deleteStudentTuitionGrade(studentID);
         boolean deleteStudentClassRel = StudentClassDAO.deleteStudentClassRel(studentID);
